+++ SourceCode/Data/User.cs
@@ -21,6 +21,6 @@ namespace ModulesRegistry.Data
         public bool IsApiAccessPermitted { get; set; }
         public string HashedPassword { get; set; }
         public int PasswordResetAttempts { get; set; }
-
+        public bool MayUploadSkpDrawing { get; set; }
     }
 }